apply_N to expression.py for being available for Expression.numerify
